move component tests to component subdirectory
