+++ src/components/ImageView/ImageView.test.tsx
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react'
 import { strict as assert } from 'assert'
-import React from 'react'
 import { describe, expect, it, vi } from 'vitest'
-import { ImageView, getHyperparamSource } from '../../src/index.js'
+import { getHyperparamSource } from '../../lib/sources/index.js'
+import ImageView from './ImageView.js'
 
 globalThis.fetch = vi.fn()
