+++ src/components/Json/Json.test.tsx
@@ -1,7 +1,6 @@
 import { fireEvent, render } from '@testing-library/react'
-import React from 'react'
 import { describe, expect, it } from 'vitest'
-import Json from '../../src/components/Json/Json.js'
+import Json from './Json.js'
 
 describe('Json Component', () => {
   it('renders primitive types correctly', () => {
