+++ src/components/SlidePanel/SlidePanel.test.tsx
@@ -1,9 +1,8 @@
 
 import { act, fireEvent, render } from '@testing-library/react'
-import React from 'react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import SlidePanel from '../../src/components/SlidePanel/SlidePanel.js'
-import { ConfigProvider } from '../../src/hooks/useConfig.js'
+import { ConfigProvider } from '../../hooks/useConfig.js'
+import SlidePanel from './SlidePanel.js'
 
 describe('SlidePanel', () => {
   // Minimal localStorage mock
