+++ src/components/Welcome/Welcome.test.tsx
@@ -1,7 +1,6 @@
 import { fireEvent, render } from '@testing-library/react'
-import React from 'react'
 import { describe, expect, it, vi } from 'vitest'
-import Welcome from '../../src/components/Welcome/Welcome.js'
+import Welcome from './Welcome.js'
 
 describe('Welcome Component', () => {
   it('renders welcome content', () => {
